lists.h fixed

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -10,7 +10,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	unsigned long int c = 0;
+	size_t c = 0;
 
 	while (h)
 	{
